wr-8: * create shared auth page components
