+++ frontend/src/libs/components/button/button.tsx
@@ -21,3 +21,4 @@ const Button: React.FC<Properties> = ({
 );
 
 export { Button };
+export { type Properties as ButtonProperties };
