+++ frontend/src/libs/components/button/styles.module.scss
@@ -3,6 +3,7 @@
   font-size: 16px;
   font-family: inherit;
   line-height: 19px;
+  border: none;
   border-radius: 7px;
   cursor: pointer;
 }
