+++ frontend/src/pages/auth/auth.tsx
@@ -1,7 +1,6 @@
 import { AppRoute } from '~/libs/enums/enums.js';
 import {
   useAppDispatch,
-  useAppSelector,
   useCallback,
   useLocation,
 } from '~/libs/hooks/hooks.js';
@@ -15,9 +14,6 @@ import { AuthLayout, SignInForm, SignUpForm } from './components/components.js';
 
 const Auth: React.FC = () => {
   const dispatch = useAppDispatch();
-  const { dataStatus } = useAppSelector(({ auth }) => ({
-    dataStatus: auth.dataStatus,
-  }));
   const { pathname } = useLocation();
 
   const handleSignInSubmit = useCallback(
@@ -47,12 +43,7 @@ const Auth: React.FC = () => {
     return null;
   };
 
-  return (
-    <AuthLayout>
-      state: {dataStatus}
-      {getScreen(pathname)}
-    </AuthLayout>
-  );
+  return <AuthLayout>{getScreen(pathname)}</AuthLayout>;
 };
 
 export { Auth };
