+++ frontend/src/pages/auth/components/auth-submit-button/styles.module.scss
@@ -1,0 +1,7 @@
+.authSubmitButton {
+  width: 100%;
+  height: 45px;
+  color: var(--white);
+  background: var(--dark-green);
+  border-radius: 7px;
+}
