+++ frontend/src/pages/auth/components/sign-in-form/sign-in-form.tsx
@@ -1,11 +1,13 @@
-import { Button, Icon, Input, Link } from '~/libs/components/components.js';
+import { Input, Link } from '~/libs/components/components.js';
 import { AppRoute } from '~/libs/enums/app-route.enum';
-import { useAppForm, useCallback, useState } from '~/libs/hooks/hooks.js';
+import { useAppForm, useCallback } from '~/libs/hooks/hooks.js';
 import {
   type UserSignInRequestDto,
   userSignInValidationSchema,
 } from '~/packages/users/users.js';
 
+import { AuthSubmitButton } from '../auth-submit-button/auth-submit-button.js';
+import { PasswordInput } from '../password-input/password-input.js';
 import { DEFAULT_LOGIN_PAYLOAD } from './libs/constants/constants.js';
 import styles from './styles.module.scss';
 
@@ -18,11 +20,6 @@ const SignInForm: React.FC<Properties> = ({ onSubmit }) => {
     defaultValues: DEFAULT_LOGIN_PAYLOAD,
     validationSchema: userSignInValidationSchema,
   });
-  const [isPasswordVisible, setPasswordVisibility] = useState(false);
-
-  const handleTogglePasswordVisibility = useCallback((): void => {
-    setPasswordVisibility((previousValue) => !previousValue);
-  }, [setPasswordVisibility]);
 
   const handleFormSubmit = useCallback(
     (event_: React.BaseSyntheticEvent): void => {
@@ -32,8 +29,7 @@ const SignInForm: React.FC<Properties> = ({ onSubmit }) => {
   );
 
   return (
-    <div className={styles.wrapper}>
-      <h2 className={styles.title}>Hello!</h2>
+    <>
       <form
         className={styles.form}
         name="loginForm"
@@ -48,36 +44,23 @@ const SignInForm: React.FC<Properties> = ({ onSubmit }) => {
             control={control}
             errors={errors}
           />
-          <div className={styles.passwordInputWrapper}>
-            <button
-              type="button"
-              className={styles.passwordToggle}
-              onClick={handleTogglePasswordVisibility}
-            >
-              <Icon
-                iconName={isPasswordVisible ? 'view' : 'hide'}
-                className={styles.icon}
-              />
-            </button>
-            <Input
-              name="password"
-              type={isPasswordVisible ? 'text' : 'password'}
-              placeholder="Enter password"
-              label="Password"
-              control={control}
-              errors={errors}
-            />
-          </div>
+          <PasswordInput
+            name="password"
+            label="Password"
+            errors={errors}
+            control={control}
+            placeholder="Enter password"
+          />
         </fieldset>
-        <Button className={styles.submitButton} type="submit" label="Sign In" />
+        <AuthSubmitButton label="Sign In" />
       </form>
-      <div>
+      <div className={styles.messageWrapper}>
         <span className={styles.message}>No account?</span>
         <Link className={styles.link} to={AppRoute.SIGN_UP}>
           Sign Up
         </Link>
       </div>
-    </div>
+    </>
   );
 };
