+++ frontend/src/pages/auth/components/sign-in-form/styles.module.scss
@@ -2,33 +2,17 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  width: 327px;
-  margin: auto;
+  width: 100%;
+  max-width: 327px;
 }
 
-.title {
-  align-self: flex-start;
-  margin: 0 0 44px;
-  color: var(--light-green);
-  font-size: 24px;
-  font-family: "Libre Baskerville", sans-serif;
+.messageWrapper {
+  position: absolute;
+  bottom: 40px;
 }
 
 .form {
   width: 100%;
-  margin-bottom: 18px;
-}
-
-.submitButton {
-  width: 100%;
-  height: 45px;
-  color: var(--white);
-  background: var(--dark-green);
-  border-radius: 7px;
-}
-
-.passwordToggle {
-  display: contents;
 }
 
 .fieldset {
@@ -38,21 +22,6 @@
   margin: 0 0 18px;
   padding: 0;
   border: none;
-}
-
-.passwordInputWrapper {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-}
-
-.icon {
-  position: absolute;
-  top: 40px;
-  right: 20px;
-  width: 16px;
-  height: 16px;
-  cursor: pointer;
 }
 
 .message {
